edge_agent/requirements.txt
bleak==0.14.3
pyyaml>=5.4
requests>=2.21.0
